requirements.txt
discord.py==2.3.0
selenium==4.15.0
webdriver-manager==4.0.1